+++ server/src/services/llmchain.py
@@ -3,7 +3,9 @@ from langchain.prompts import PromptTemplate
 from langchain.schema import StrOutputParser
 from langchain_core.runnables import RunnablePassthrough
 
-API_KEY = "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"
+from ..config.settings import settings
+
+API_KEY = settings.llm_api_key
 llm_id = "databricks/dolly-v2-3b"
 transformer_id = "sentence-transformers/all-MiniLM-L6-v2"
 template = """Use the following pieces of context to answer the question at the end. If you don't know the answer, just say that you don't know, don't try to make up an answer. Use three sentences maximum. Keep the answer as concise as possible.
